giao diện trang chủ và chi tiết sản phẩm

+++ src/main/java/com/smashvn/shop/entity/DanhGia.java
@@ -27,7 +27,7 @@ public class DanhGia {
     private SanPham sanPham;
 
     @Column(name = "so_sao", nullable = false)
-    private Integer soSao;
+    private Double soSao;
 
     @Column(name = "binh_luan", columnDefinition = "NVARCHAR(MAX)")
     private String binhLuan;
